tests/wpt/web-platform-tests/clipboard-apis/async-write-text-read-text-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Async Clipboard writeText -> readText tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async (t) => {
  const input = "Clipboard writeText -> readText test data";

  await navigator.clipboard.writeText(input);
  const output = await navigator.clipboard.readText();

  assert_equals(output, input);
}, "Verify write and read clipboard (DOMString)");
</script>
Note: This is a manual test because it writes/reads to the shared system
clipboard and thus cannot be run async with other tests that might interact
with the clipboard.
